Beep.Python.Runtime.Host/PythonServers/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
pywin32>=306; sys_platform == 'win32'
numpy>=1.24.0
